Handle reactivity differently for the deltas.
First we create the workflowTree when the component is created, using the Vuex store getter
(this can be changed later too! moved out of the store!).

Then we add a local copy of the workflows from the Vuex store, and set up a Watcher. Any
time the workflows change in the store, we are notified (or any attribute of the
workflow - deep: true).

When the workflows change, then we know that we need to re-create the tree. A TODO is
to stop re-creating the tree, and instead implement a per-component strategy. So the Graph
will update its edges and nodes. The tree will update each tree item. And so it goes.