Update treeitem children elements.
When the infinite tree was first implemented, the structure of the tree was different than
what we have now.

That's because we had several changes to the tree, such as fixing element width and how
the elements react when the user re-sizes the browser (overflow, word-break, etc).

We also had improvements to how the elements were displayed, how many events were
triggered, and other changes for the deltas.

This caused this pull request to actually break some of these changes/improvements.
This commit should fix that, though tests may still need updating.

But the tree functionality is back to what it was, plus the infinite tree.