Handle deltas subscriptions according to the Vue route selected.
When the user navigates to the view, we will have a Tree view, and also a
deltas subscription.

If the user adds more tree views, we will keep using a single subscription.

If the user removes all tree views, we will stop the subscription.

If the user leaves the page, we will destroy the subscription.

Finally, if the user changes the workflow, we destroy the subscription, and
create a new one, also adding a new tree view as that is the default.