Update dependencies
I used `ncu` to get a list of dependencies that had updates available. Then went through these dependencies individually, updating each one and checking if the build was working OK.

`vuetify` was not updated as it is being updated in #466.

Updating `eslint-config-vuetify` resulted in some weird errors in `yarn install`:

error An unexpected error occurred: "could not find a copy of eslint to link in /home/kinow/Development/python/workspace/cylc-ui/node_modules/eslint-config-vuetify/node_modules".

So I reverted it and will take care of that dependency later (it is a dev dependency).