Create Subscriptions for the views, that are organized by Query name in the Workflow Service.
This way we can have multiple components and views re-using the same queries
that are merged. Or they can use different queries.

When different queries are used (with different names, of course), the subscriptions
for these queries are kept apart. This is useful when you have, for instance, a subscription
like GScan's that has no variables, while Tree's query uses variables and different
structure.

The lifecycle of subscriptions is not maintained  by the Workflow Service. It is now maintained
via lifecycle methods of components and navigation guards of views. The code for that is in
mixins.

After these changes, the Workflow view can now display any View or Component that uses Subscriptions.
And the Workflow view does not have to coordinate when widgets are created with their subscriptions.

Unit tests and e2e tests will have to be updated, and new tests added as well. But with these
changes, the Table view PR should be more easily merged. Besides fixing tests, we will also need
to sort out the query merge that appears to have a few cases where queries are not merged
correctly.