Began update to not share options betwen workflows
The work done on the branch so far makes the options state persist
between workflows, which is not desired. This commit is an attempt to
fix this by having an array of options (optionsArray) with an options
object for each workflow. However I am struggling with the persisted
state caused by reactivity (I think) in the options object.